use complex interpolation in read_tide_model

+++ pyTMD/read_tide_model.py
@@ -343,8 +343,8 @@ def bilinear_interp(ilon,ilat,idata,lon,lat):
 	#-- Convert output data coordinates to radians
 	xphi = lon*dtr
 	xth = (90.0 - lat)*dtr
-	#-- interpolate gridded ur values to data
-	data = np.zeros_like(lon)
+	#-- interpolate gridded data values to data
+	data = np.zeros_like(lon,dtype=np.complex128)
 	for i,l in enumerate(lon):
 		#-- calculating the indices for the original grid
 		dx = (ilon - np.floor(lon[i]/dlon)*dlon)**2
